agent use generic config
